Moved examples to JDK11 run as file
