+++ src/examples/ClockWatcher.java
@@ -13,7 +13,6 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.oddcyb.microbots.examples;
 
 import java.util.concurrent.Executors;
 import java.util.concurrent.ScheduledExecutorService;
