+++ src/examples/Greeter.java
@@ -13,7 +13,6 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.oddcyb.microbots.examples;
 
 import java.util.Random;
 import org.oddcyb.microbots.Robot;
